+++ src/outliner/views/script_obj.py
@@ -1,8 +1,8 @@
 import ast
+import sys
 
 from importlib import util
 from pathlib import Path
-from functools import partial
 
 
 class Script_Obj:
@@ -19,10 +19,13 @@ class Script_Obj:
     ):
         self.script_path = script_path
         self.object_ast = self._load_file_ast()
+        self._imported_modules = self._find_imported_modules()
 
-    def _find_imported_modules(self) -> list:
+        self._add_imports()
+
+    def _find_imported_modules(self) -> iter:
         """
-        A generator with all imports the object contains in its namespace
+        A generator with all imports the ast object contains in its namespace
         """
         for node in ast.walk(self.object_ast):
             if isinstance(node, ast.Import):
@@ -37,11 +40,11 @@ class Script_Obj:
         with open(self.script_path, "r") as file:
             return ast.parse(file.read())
 
-    def add_imports(self) -> None:
+    def _add_imports(self) -> None:
         """
         Add all found imports to the running gloabal namespace
         """
-        ...
+        sys.path.append(i for i in self._imported_module)
 
     def module(self):
         """
@@ -53,41 +56,3 @@ class Script_Obj:
         moduleSpec.loader.exec_module(module_obj)
 
         return module_obj
-
-
-class Trace_Obj:
-    def __init__(self, script_obj: Script_Obj, obj_invoking: str):
-        self.obj_invoking = obj_invoking
-        self.script_obj = script_obj
-
-    def _get_object_arguments(self, obj: str):
-        parenthesis_1 = obj.index("(")
-        parenthesis_2 = obj.index(")")
-        obj_arguments = obj[parenthesis_1 + 1 : parenthesis_2].split(",")
-        return obj_arguments if any(obj_arguments) else None
-
-    def _create_obj_instance(self, namespace: any, object_: str):
-        object_name = object_.split("(")[0]
-        object_arguments = self._get_object_arguments(object_)
-
-        obj_instance = getattr(namespace, object_name, None)
-
-        if object_arguments is not None:
-            obj_instance = partial(obj_instance, *object_arguments)
-
-        return obj_instance
-
-    def create(self):
-        """Returns a instance of the object"""
-
-        running_obj = None
-        father = self.script_obj.module()
-        objs = self.obj_invoking.split(".") or [self.obj_invoking]
-
-        for number, _ in enumerate(objs):
-            running_obj = self._create_obj_instance(father, objs[number])
-            if len(objs) == number:
-                break
-            father = running_obj()
-
-        return running_obj
